requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=9.0.0
pandas>=1.5.0
numpy>=1.23.0
datasets>=2.0.0
tqdm>=4.0.0
